Show Site Header on Home page
Completed the conditional to allow for the home page to have the same site header. I lack information as to what was supposed to be put there otherwise, so for now it seems redundant and we could get rid of the conditional altogether, however I'll leave it in case there was a different reason to include it as such.